fix: increase rate limit fior addTeam
