--- packages/server/graphql/types/AddOrgPayload.ts
@@ -1,18 +1,0 @@
-import {GraphQLObjectType} from 'graphql'
-import type {GQLContext} from '../graphql'
-import {resolveOrganization} from '../resolvers'
-import {addTeamFields} from './AddTeamPayload'
-import Organization from './Organization'
-
-const AddOrgPayload = new GraphQLObjectType<any, GQLContext>({
-  name: 'AddOrgPayload',
-  fields: () => ({
-    organization: {
-      type: Organization,
-      resolve: resolveOrganization
-    },
-    ...addTeamFields
-  })
-})
-
-export default AddOrgPayload
